refactor: firebase 관련 코드를 projects.js에서 분리한다
