+++ projects/index.html
@@ -792,7 +792,8 @@
 
     <script src="../js/i18n.js"></script>
     <script src="../js/main.js"></script>
-    <script type="module" src="projects.js"></script>
+    <script src="projects.js"></script>
+    <script type="module" src="firebase.js"></script>
 </body>
 
 </html>
